% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas.R
\name{ctwas}
\alias{ctwas}
\title{ctwas: Function to perform SuSiE-based causal tissue-gene pair and direct causal variant estimation}
\usage{
ctwas(
  by,
  bXest,
  LD,
  Noutcome,
  L_causal = 10,
  eigen_thres = 0.999,
  pip_thres_cred = 0.95,
  susie_iter = 500
)
}
\description{
This function performs the SuSiE (Sum of Single Effects) algorithm to estimate the causal tissue-gene pairs and direct causal variants using a provided set of summary statistics and linkage disequilibrium (LD) matrix. It returns a list of estimates and their associated statistics, including credible sets and Pratt estimations.
}
