% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgfm.R
\name{tgfm}
\alias{tgfm}
\title{tgfm: Function to perform eQTL fine-mapping and causal tissue-gene pair estimation}
\usage{
tgfm(
  by,
  bX,
  LD,
  Nvec,
  L.eqtl = 5,
  L_causal = 10,
  eigen_thres = 0.999,
  pip_thres_cred = 0.5,
  eqtl_sampling_time = 100,
  causal_sampling_time = 100,
  eqtl_thres = 0.05,
  susie_iter = 1000
)
}
\description{
This function performs fine-mapping of eQTLs and estimates causal tissue-gene pairs using summary statistics, linkage disequilibrium (LD) matrix, and resampling methods. It returns a list of estimates and associated statistics, including credible sets, PIP resampling, and Pratt estimations.
}
